ownership/requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
cachetools==4.1.0         # via google-auth
certifi==2020.4.5.2       # via requests
cffi==1.14.0              # via cryptography
chardet==3.0.4            # via requests
click==7.1.2              # via -r requirements/base.in
cryptography==2.9.2       # via jwcrypto
github3.py==1.3.0         # via -r requirements/base.in
google-auth-oauthlib==0.4.1  # via gspread
google-auth==1.16.1       # via google-auth-oauthlib, gspread
gspread==3.6.0            # via -r requirements/base.in
idna==2.9                 # via requests
jwcrypto==0.7             # via github3.py
oauthlib==3.1.0           # via requests-oauthlib
pyasn1-modules==0.2.8     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, rsa
pycparser==2.20           # via cffi
python-dateutil==2.8.1    # via github3.py
requests-oauthlib==1.3.0  # via google-auth-oauthlib
requests==2.23.0          # via github3.py, gspread, requests-oauthlib
rsa==4.0                  # via google-auth
six==1.15.0               # via cryptography, google-auth, python-dateutil
uritemplate==3.0.1        # via github3.py
urllib3==1.25.9           # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
